requirements.txt
absl-py==1.0.0
aiohttp==3.8.1  
aiosignal==1.2.0
anyio==3.4.0
astunparse==1.6.3
async-timeout==4.0.1
attrs==21.2.0
autopep8==1.6.0
backoff==1.11.1
cachetools==4.2.4
certifi==2021.10.8
cffi==1.15.0
chardet==3.0.4
charset-normalizer==2.0.8
cryptography==35.0.0
Faker==9.9.0
flatbuffers==2.0
frozenlist==1.2.0
future==0.18.2
gast==0.4.0
gcloud-aio-auth==3.7.0
gcloud-aio-pubsub==4.4.0
google-api-core==2.2.2
google-auth==2.3.3
google-auth-oauthlib==0.4.6
google-cloud-pubsub==2.9.0
google-pasta==0.2.0
googleapis-common-protos==1.53.0
grpc-google-iam-v1==0.12.3
grpcio==1.42.0
grpcio-status==1.42.0
h11==0.12.0
h5py==3.6.0
httpcore==0.14.3
httpx==0.21.1
idna==3.3
importlib-metadata==4.8.2
kafka==1.3.5
kafka-python==2.0.2
keras==2.7.0
Keras-Preprocessing==1.1.2
libclang==12.0.0
libcst==0.3.23
Markdown==3.3.6
multidict==5.2.0
mypy-extensions==0.4.3
numpy==1.21.4
oauthlib==3.1.1
opt-einsum==3.3.0
pandas==1.3.4
proto-plus==1.19.8
protobuf==3.19.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.8.0
pycparser==2.21
PyJWT==2.3.0
python-dateutil==2.8.2
pytz==2021.3
PyYAML==6.0
requests==2.26.0
requests-oauthlib==1.3.0
rfc3986==1.5.0
rsa==4.8
six==1.16.0
sniffio==1.2.0
tensorboard==2.7.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tensorflow==2.7.0
tensorflow-estimator==2.7.0
tensorflow-io-gcs-filesystem==0.22.0
termcolor==1.1.0
text-unidecode==1.3
toml==0.10.2
typing-extensions==4.0.1
typing-inspect==0.7.1
urllib3==1.26.7
Werkzeug==2.0.2
wrapt==1.13.3
yarl==1.7.2
zipp==3.6.0